pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "djangocms-form-builder-countries"
version = "0.1.0"
description = "Country selector field plugin for djangocms-form-builder"
readme = "README.rst"
license = {text = "MIT"}
authors = [
    {name = "Daniel Altiparmak"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Framework :: Django CMS",
    "Framework :: Django CMS :: 4.1",
    "Framework :: Django CMS :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
keywords = [
    "django",
    "django-cms",
    "djangocms-form-builder",
    "countries",
    "form",
    "plugin",
]
dependencies = [
    "Django>=4.2",
    "django-cms>=4.1",
    "djangocms-form-builder>=2.0",
    "django-countries>=7.0",
]

[project.urls]
Homepage = "https://github.com/daltiparmak/djangocms-form-builder-countries"
Documentation = "https://github.com/daltiparmak/djangocms-form-builder-countries#readme"
Repository = "https://github.com/daltiparmak/djangocms-form-builder-countries"
Issues = "https://github.com/daltiparmak/djangocms-form-builder-countries/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "pytest-cov>=4.0",
    "tox>=4.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["djangocms_form_builder_countries*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["djangocms_form_builder_countries"]
branch = true
omit = [
    "*/migrations/*",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
show_missing = true
fail_under = 70

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["djangocms_form_builder_countries"]
